RKBcoin 罗克币
RKcoin encrypted digital currency Scrypt mantra - Eork algorithm, designed according to the concept of bitcoins integrated algorithms, the total is: 180 million, referred to as: ByarieScroyptHash285, purse with six super hash algorithm, communication port 9332, RPC - port 9562, is a function of the numerical calculation based on field research and development of new P2P encrypted e-money.
RKcoin is entering the international mainstream digital currency in circulation, the current market transactions in Europe and other countries. Not by any central authority management, fully decentralized, stock completely on characteristics of the algorithm to determine. In view of the institutional investors, RKcoin on anonymously, without borders, not easy to crack and secondary more stringent.